Update and rename 2021-10-24-Example-post.md to 2021-10-23-Example-post.md
After visiting rossetti Archive, I would like to talk briefly about my feelings. First, its pages are classical and visually aesthetically pleasing. Secondly, these works have been digitized and uploaded digitally, so that they retain the classical flavor to a certain extent and can accurately convey information. At the same time, the museum displays a variety of exhibits and provides logical information, such as alphabetical and chronological sorting, which is easy to browse and refer to. However, in terms of content, it may not be that rich, which is one of the limitations, but I have to say that its timeline display is very good. This kind of digital technology can save a lot of important data information to a large extent, and it is easy to find and reuse, I think it is very meaningful.